requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
aniemore==1.2.3
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.0
decorator==5.1.1
dill==0.3.8
ffmpeg-python==0.2.0
filelock==3.13.4
Flask==3.0.3
flask-sock==0.7.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.22.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
keras==3.2.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai-whisper==20230124
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
simple-websocket==1.0.0
six==1.16.0
soundfile==0.12.1
sox==1.5.0
soxr==0.3.7
srt==3.5.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
tokenizers==0.13.3
torch==2.2.2
torchaudio==2.2.2
tqdm==4.66.2
transformers==4.26.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vosk==0.3.44
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
